Begin CLI refactor
