--- elk/train_evaluate.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-python -m elk.train suspicious-wu
-# python -m elk.evaluate deberta-v2-xxlarge-mnli boolq --prefix normal --num-data 1000
