+++ elk/training/train.py
@@ -4,14 +4,14 @@ import random
 
 import numpy as np
 import torch
-from argparse import ArgumentParser
 from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import accuracy_score, roc_auc_score
 from tqdm.auto import tqdm
 
-from .files import elk_cache_dir
-from .training.ccs import CCS
-from .training.preprocessing import load_hidden_states
+from ..files import elk_cache_dir
+from .ccs import CCS
+from .parser import get_training_parser
+from .preprocessing import load_hidden_states
 
 
 @torch.autocast("cuda", enabled=torch.cuda.is_available())
@@ -97,52 +97,14 @@ def train(args):
 
     ccs_models.reverse()
     lr_models.reverse()
-    return lr_models, ccs_models
-
-
-if __name__ == "__main__":
-    parser = ArgumentParser()
-    parser.add_argument("name", type=str, help="Name of the experiment")
-    parser.add_argument(
-        "--device",
-        type=str,
-        help="PyTorch device to use. Default is cuda:0 if available.",
-    )
-    parser.add_argument(
-        "--loss",
-        type=str,
-        default="squared",
-        choices=("js", "squared"),
-        help="Loss function used for CCS.",
-    )
-    parser.add_argument(
-        "--optimizer",
-        type=str,
-        default="lbfgs",
-        choices=("adam", "lbfgs"),
-        help="Optimizer for CCS. Should be adam or lbfgs.",
-    )
-    parser.add_argument("--seed", type=int, default=0)
-    parser.add_argument(
-        "--weight-decay",
-        type=float,
-        default=0.01,
-        help=(
-            "Weight decay for CCS when using Adam. Used as L2 regularization in LBFGS."
-        ),
-    )
-    args = parser.parse_args()
-
-    # Default to CUDA iff available
-    if args.device is None:
-        import torch
-
-        args.device = "cuda" if torch.cuda.is_available() else "cpu"
-
-    lr_models, ccs_models = train(args)
 
     path = elk_cache_dir() / args.name
     with open(path / "lr_models.pkl", "wb") as file:
         pickle.dump(lr_models, file)
 
     torch.save(ccs_models, path / "ccs_models.pt")
+
+
+if __name__ == "__main__":
+    args = get_training_parser().parse_args()
+    train(args)
