+++ pyproject.toml
@@ -32,5 +32,8 @@ dev = [
     "pytest",
 ]
 
+[project.scripts]
+elk = "elk.__main__:run"
+
 [tool.setuptools]
 packages = ["elk"]
